cleanup: renamed multichain to unichain, added unichain deploy script
