+++ packages/data-union-solidity/package.json
@@ -16,8 +16,9 @@
   "author": "Streamr Network AG",
   "license": "AGPL-3.0",
   "devDependencies": {
-    "@nomiclabs/hardhat-waffle": "^2.0.1",
     "@nomiclabs/hardhat-ethers": "^2.0.3",
+    "@nomiclabs/hardhat-waffle": "^2.0.1",
+    "@streamr/config": "^0.0.3",
     "@typechain/ethers-v5": "^6.0.0",
     "@types/chai": "^4.2.22",
     "@types/debug": "^4.1.7",
